renaming element/html modules
